% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F3B-package.R
\docType{package}
\name{F3B-package}
\alias{F3B-package}
\title{Model for Krill Fishery Management Strategy Evaluation}
\description{
F3B
}
\details{
F3B package
}
\author{
George Watters \email{george.watters@noaa.gov}

Sam Woodman \email{sam.woodman@noaa.gov}
}
\keyword{package}
