% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{F3B}
\alias{F3B}
\title{A model for within-year biomass dynamics in three nested biomass pools - Z within Y within X}
\usage{
F3B(t, state, parameters, survey, surveytime, catfuncX, catfuncY, catfuncZ)
}
\arguments{
\item{t}{todo}

\item{state}{todo}

\item{parameters}{todo}

\item{survey}{todo}

\item{surveytime}{todo}

\item{catfuncX}{todo}

\item{catfuncY}{todo}

\item{catfuncZ}{todo}
}
\value{
A list of ...
}
\description{
A model for within-year biomass dynamics in three nested biomass pools - Z within Y within X
}
\details{
Run the model...
}
