% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxillary.R
\name{GetCVp}
\alias{GetCVp}
\title{Calculate ...}
\usage{
GetCVp(XYZ.mean = 19158000, XYZ.CV = 0.11, YZ.mean = 6853000, YZ.CV = 0.16)
}
\arguments{
\item{XYZ.mean}{todo}

\item{XYZ.CV}{todo}

\item{YZ.mean}{todo}

\item{YZ.CV}{todo}
}
\description{
Calculate ...
}
\details{
todo
}
