% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRAconsumption.R
\name{GetRAconsumption}
\alias{GetRAconsumption}
\title{todo}
\usage{
GetRAconsumption()
}
\description{
calculate predator demand density in tonnes for all of summer?
}
\details{
code and data files used here provided by Vicky Warwick-Evans to WG-EMM e-group or to GW via email
Todo: move paths to be function arguments
}
