% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_parallel.R
\name{MC.MSEwithF3B.par}
\alias{MC.MSEwithF3B.par}
\title{A wrapper to do many Monte Carlo simulations - the parallel version}
\usage{
MC.MSEwithF3B.par(
  nsims = 3,
  MC.rates,
  MC.imports,
  MC.resets,
  MC.fishing,
  suppress = FALSE,
  num.cores = NULL,
  parallel.seed = NULL
)
}
\arguments{
\item{nsims}{number of Monte Carlo simulations to run. The default is 3}

\item{MC.rates}{todo}

\item{MC.imports}{todo}

\item{MC.resets}{todo}

\item{MC.fishing}{todo}

\item{suppress}{logical; if not running in parallel,
should the function print which simulation it is currently running.
Default is FALSE}

\item{num.cores}{Number of CPUs to over which to distribute computations.
Defaults to \code{NULL}, which uses one fewer than the number of cores
reported by \code{\link[parallel]{detectCores}}}

\item{parallel.seed}{Should be \code{NULL}; output of \code{\link{parallel::clusterSetRNGStream}}}
}
\value{
A three dimensional array...
}
\description{
A wrapper to do many Monte Carlo simulations - the parallel version
}
\details{
todo
todo: likely rename inputs so they're the same as MSEwithF3B for consistency
}
