% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_parallel.R
\name{MSEwithF3B.wrap}
\alias{MSEwithF3B.wrap}
\title{Wrapper function}
\usage{
MSEwithF3B.wrap(i, ...)
}
\arguments{
\item{i}{index of the MC simulation; ignored}

\item{...}{passed to \code{\link{MSEwithF3B}}}
}
\description{
Wrapper function for \code{\link{MSEwithF3B}}
}
\details{
Wrapper for \code{\link{MSEwithF3B}}. The purpose of this function is to allow
\code{X} from \code{\link[parallel:clusterApply]{parLapply}}
(in this case, the index fo the current simulation) to be
passed to \code{\link{MSEwithF3B}}, and ignored
}
