% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxillary.R
\name{ObjectiveF3B}
\alias{ObjectiveF3B}
\title{Calculate ...}
\usage{
ObjectiveF3B(x, rates, resets, fishing)
}
\arguments{
\item{x}{todo}

\item{rates}{todo}

\item{resets}{todo}

\item{fishing}{tooo}
}
\description{
Calculate ...
}
\details{
todo
}
