% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ss}
\alias{plot.ss}
\title{Plot F3B model outputs}
\usage{
\method{plot}{ss}(indata, pick.me = c("Y", "Z", "survey"))
}
\arguments{
\item{indata}{output of...}

\item{pick.me}{a character vector; one or more of "Y","Z", or "survey"}
}
\description{
Plot F3B model outputs
}
\details{
Plot the weekly biomass time series, along with survey estimates and effective harvest rates
}
