% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{PositiveF3B}
\alias{PositiveF3B}
\title{An event function that tries to ensure biomasses >= 0}
\usage{
PositiveF3B(
  t,
  state,
  parameters,
  survey,
  surveytime,
  catfuncX,
  catfuncY,
  catfuncZ
)
}
\arguments{
\item{t}{todo}

\item{state}{todo}

\item{parameters}{todo}

\item{survey}{todo}

\item{surveytime}{todo}

\item{catfuncX}{todo}

\item{catfuncY}{todo}

\item{catfuncZ}{todo}
}
\description{
An event function that tries to ensure biomasses >= 0
}
