% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxillary.R
\name{TuneF3B}
\alias{TuneF3B}
\title{Calculate ...}
\usage{
TuneF3B(rates, resets, fishing)
}
\arguments{
\item{rates}{todo}

\item{resets}{todo}

\item{fishing}{tooo}
}
\value{
A list...
}
\description{
Calculate ...
}
\details{
todo
}
