% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{MSEwithF3B}
\alias{MSEwithF3B}
\title{Main function for running F3B}
\usage{
MSEwithF3B(rates, imports, resets, fishing)
}
\arguments{
\item{rates}{todo}

\item{imports}{todo}

\item{resets}{todo}

\item{fishing}{todo}
}
\description{
Piece everything together into the main function
}
